+++ pkg/apiserver/server.go
@@ -5,29 +5,12 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/gorilla/mux"
 	log "github.com/sirupsen/logrus"
 )
 
 //Server -
 func Server(port int) error {
 	e := setAPIEndpoints()
-	// Create a new router
-	router := mux.NewRouter()
-
-	// Define the retrieval endpoints for Plunder Server configuration
-	router.HandleFunc("/config", getConfig).Methods("GET")
-
-	// Define the creation endpoints for Plunder Server Configuration
-	router.HandleFunc("/config", postConfig).Methods("POST")
-
-	// Define the retrieval endpoints for Plunder Deployment configuration
-	router.HandleFunc("/deployment", getDeployment).Methods("GET")
-
-	// Define the creation and modification endpoints for Plunder Deployment configuration
-	router.HandleFunc("/deployment", postDeployment).Methods("POST")
-	router.HandleFunc("/deployment/{ID}", nil).Methods("POST")
-	router.HandleFunc("/deployment/{ID}", nil).Methods("DELETE")
 	log.Infof("Starting API server on port %d", port)
 
 	err := http.ListenAndServe(fmt.Sprintf(":%d", port), e)
